api/requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.2
Flask-Cors==3.0.10
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
packaging==23.1
protobuf==4.21.12
python-dotenv==1.0.0
six==1.16.0
Werkzeug==2.3.4
